Fixing pep8

+++ project/helper.py
@@ -8,6 +8,7 @@ from mangrove.form_model.validation import NumericConstraint, TextConstraint
 from mangrove.utils.helpers import slugify
 from mangrove.utils.types import is_empty, is_sequence, is_not_empty
 
+
 def get_or_create_data_dict(dbm, name, slug, primitive_type, description=None):
     try:
         #  Check if is existing
@@ -19,7 +20,7 @@ def get_or_create_data_dict(dbm, name, slug, primitive_type, description=None):
     return ddtype
 
 
-def create_question(post_dict,dbm=None):
+def create_question(post_dict, dbm=None):
     if dbm is None:
         dbm = get_db_manager()
 
@@ -31,13 +32,13 @@ def create_question(post_dict,dbm=None):
     else:
         datadict_slug = str(slugify(unicode(post_dict.get('title'))))
 
-    ddtype = get_or_create_data_dict(dbm = dbm, name=post_dict.get('code'), slug=datadict_slug,
+    ddtype = get_or_create_data_dict(dbm=dbm, name=post_dict.get('code'), slug=datadict_slug,
                                primitive_type=post_dict.get('type'), description=post_dict.get('title'))
 
     if post_dict["type"] == "text":
-        return _create_text_question(post_dict,ddtype)
+        return _create_text_question(post_dict, ddtype)
     if post_dict["type"] == "integer":
-        return _create_integer_question(post_dict,ddtype)
+        return _create_integer_question(post_dict, ddtype)
     if post_dict["type"] == "select":
         return _create_select_question(post_dict, single_select_flag=False, ddtype=ddtype)
     if post_dict["type"] == "date":
@@ -47,9 +48,9 @@ def create_question(post_dict,dbm=None):
 
 
 def create_questionnaire(post, dbm=get_db_manager()):
-    entity_data_dict_type = get_or_create_data_dict(dbm = dbm, name="eid", slug="entity_id",primitive_type="string", description="Entity ID")
+    entity_data_dict_type = get_or_create_data_dict(dbm=dbm, name="eid", slug="entity_id", primitive_type="string", description="Entity ID")
     entity_id_question = TextField(name="What are you reporting on?", question_code="eid", label="Entity being reported on",
-                                   entity_question_flag=True,ddtype = entity_data_dict_type,length=TextConstraint(min=1, max=12))
+                                   entity_question_flag=True, ddtype=entity_data_dict_type, length=TextConstraint(min=1, max=12))
     return FormModel(dbm, entity_type=post["entity_type"], name=post["name"], fields=[entity_id_question], form_code='default', type='survey')
 
 
@@ -68,17 +69,17 @@ def get_code_and_title(fields):
     return [(each_field.question_code, each_field.name)for each_field in fields]
 
 
-def _create_text_question(post_dict,ddtype):
+def _create_text_question(post_dict, ddtype):
     max_length_from_post = post_dict.get("max_length")
     min_length_from_post = post_dict.get("min_length")
     max_length = max_length_from_post if not is_empty(max_length_from_post) else None
     min_length = min_length_from_post if not is_empty(min_length_from_post) else None
     length = TextConstraint(min=min_length, max=max_length)
     return TextField(name=post_dict["title"], question_code=post_dict["code"].strip(), label="default",
-                     entity_question_flag=post_dict.get("is_entity_question"), length=length,ddtype= ddtype)
+                     entity_question_flag=post_dict.get("is_entity_question"), length=length, ddtype=ddtype)
 
 
-def _create_integer_question(post_dict,ddtype):
+def _create_integer_question(post_dict, ddtype):
     max_range_from_post = post_dict["range_max"]
     max_range = max_range_from_post if not is_empty(max_range_from_post) else None
     range = NumericConstraint(min=post_dict["range_min"], max=max_range)
@@ -86,7 +87,7 @@ def _create_integer_question(post_dict,ddtype):
 
 
 def _create_date_question(post_dict, ddtype):
-    return DateField(name=post_dict["title"], question_code=post_dict["code"].strip(), label="default", date_format=post_dict.get('date_format'), ddtype= ddtype)
+    return DateField(name=post_dict["title"], question_code=post_dict["code"].strip(), label="default", date_format=post_dict.get('date_format'), ddtype=ddtype)
 
 
 def _create_select_question(post_dict, single_select_flag, ddtype):
@@ -102,4 +103,3 @@ def get_submissions(questions, submissions):
         assert isinstance(s, dict) and s.get('values') is not None
     formatted_list = [[each.get('created'), each.get('channel'), each.get('status'), each.get('error_message')] + [each.get('values').get(q[0]) for q in questions] for each in submissions]
     return [tuple(each) for each in formatted_list]
-
